Move various utils classes to their own package
