+++ javaparser-core/src/main/java/com/github/javaparser/ast/TypeParameter.java
@@ -21,8 +21,7 @@
  
 package com.github.javaparser.ast;
 
-import static com.github.javaparser.Position.pos;
-import static com.github.javaparser.ast.internal.Utils.ensureNotNull;
+import static com.github.javaparser.utils.Utils.ensureNotNull;
 
 import java.util.List;
