+++ javaparser-core/src/main/java/com/github/javaparser/ast/expr/FieldAccessExpr.java
@@ -28,7 +28,7 @@ import com.github.javaparser.ast.visitor.VoidVisitor;
 
 import java.util.List;
 
-import static com.github.javaparser.ast.internal.Utils.*;
+import static com.github.javaparser.utils.Utils.*;
 
 /**
  * @author Julio Vilmar Gesser
