+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/ReferenceType.java
@@ -22,7 +22,7 @@
 package com.github.javaparser.ast.type;
 
 import static com.github.javaparser.Position.pos;
-import static com.github.javaparser.ast.internal.Utils.ensureNotNull;
+import static com.github.javaparser.utils.Utils.ensureNotNull;
 
 import java.util.List;
