+++ javaparser-core/src/main/java/com/github/javaparser/ast/type/Type.java
@@ -27,7 +27,7 @@ import com.github.javaparser.ast.expr.AnnotationExpr;
 
 import java.util.List;
 
-import static com.github.javaparser.ast.internal.Utils.*;
+import static com.github.javaparser.utils.Utils.*;
 
 /**
  * @author Julio Vilmar Gesser
